Got rid of tuple_access_strategy_test_util.h
