+++ test/storage/tuple_access_strategy_test.cpp
@@ -1,11 +1,104 @@
+#include <unordered_map>
+
 #include "common/test_util.h"
-#include "storage/tuple_access_strategy_test_util.h"
 #include "common/typedefs.h"
+#include "storage/storage_utils.h"
+#include "storage/tuple_access_strategy.h"
+#include "util/storage_test_util.h"
 
 namespace terrier {
 struct TupleAccessStrategyTests : public ::testing::Test {
   storage::RawBlock *raw_block_ = nullptr;
   storage::BlockStore block_store_{1};
+
+// This does NOT return a sensible tuple in general. This is just some filler
+// to write into the storage layer and is devoid of meaning outside of this class.
+// This is distinct from a ProjectedRow and should only be used to exercise TupleAccessStrategy's ability to access
+// RawBlocks.
+// ProjectedRow has additional checks and invariants only meaningful at the DataTable level
+  struct FakeRawTuple {
+    template<typename Random>
+    FakeRawTuple(const storage::BlockLayout &layout, Random &generator)
+        : layout_(layout), attr_offsets_(), contents_(new byte[layout.tuple_size_]) {
+      uint32_t pos = 0;
+      for (uint16_t col = 0; col < layout.num_cols_; col++) {
+        attr_offsets_.push_back(pos);
+        pos += layout.attr_sizes_[col];
+      }
+      testutil::FillWithRandomBytes(layout.tuple_size_, contents_, generator);
+    }
+
+    ~FakeRawTuple() { delete[] contents_; }
+
+    // Since all fields we store in pages are equal to or shorter than 8 bytes,
+    // we can do equality checks on uint64_t always.
+    // 0 return for non-primary key indexes should be treated as null.
+    uint64_t Attribute(const storage::BlockLayout &layout, uint16_t col) {
+      return storage::ReadBytes(layout.attr_sizes_[col],
+                                contents_ + attr_offsets_[col]);
+    }
+
+    const storage::BlockLayout &layout_;
+    std::vector<uint32_t> attr_offsets_;
+    byte *contents_;
+  };
+
+// Write the given fake tuple into a block using the given access strategy,
+// at the specified offset
+  void InsertTuple(FakeRawTuple &tuple, storage::TupleAccessStrategy &tested, const storage::BlockLayout &layout,
+                   storage::TupleSlot slot) {
+    for (uint16_t col = 0; col < layout.num_cols_; col++) {
+      uint64_t col_val = tuple.Attribute(layout, col);
+      if (col_val != 0 || col == PRESENCE_COLUMN_ID)
+        storage::WriteBytes(layout.attr_sizes_[col],
+                            tuple.Attribute(layout, col),
+                            tested.AccessForceNotNull(slot, col));
+      else
+        tested.SetNull(slot, col);
+      // Otherwise leave the field as null.
+    }
+  }
+
+// Check that the written tuple is the same as the expected one
+  void CheckTupleEqual(FakeRawTuple &expected, storage::TupleAccessStrategy &tested, const storage::BlockLayout &layout,
+                       storage::TupleSlot slot) {
+    for (uint16_t col = 0; col < layout.num_cols_; col++) {
+      uint64_t expected_col = expected.Attribute(layout, col);
+      // 0 return for non-primary key indexes should be treated as null.
+      bool null = (expected_col == 0) && (col != PRESENCE_COLUMN_ID);
+      byte *col_slot = tested.AccessWithNullCheck(slot, col);
+      if (!null) {
+        EXPECT_TRUE(col_slot != nullptr);
+        EXPECT_EQ(expected.Attribute(layout, col),
+                  storage::ReadBytes(layout.attr_sizes_[col], col_slot));
+      } else {
+        EXPECT_TRUE(col_slot == nullptr);
+      }
+    }
+  }
+
+// Using the given random generator, attempts to allocate a slot and write a
+// random tuple into it. The slot and the tuple are logged in the given map.
+// Checks are performed to make sure the insertion is sensible.
+  template<typename Random>
+  std::pair<const storage::TupleSlot, FakeRawTuple> &TryInsertFakeTuple(
+      const storage::BlockLayout &layout, storage::TupleAccessStrategy &tested, storage::RawBlock *block,
+      std::unordered_map<storage::TupleSlot, FakeRawTuple> &tuples, Random &generator) {
+    storage::TupleSlot slot;
+    // There should always be enough slots.
+    EXPECT_TRUE(tested.Allocate(block, slot));
+    EXPECT_TRUE(tested.ColumnNullBitmap(block,
+                                        PRESENCE_COLUMN_ID)->Test(slot.GetOffset()));
+
+    // Construct a random tuple and associate it with the tuple slot
+    auto result =
+        tuples.emplace(std::piecewise_construct, std::forward_as_tuple(slot), std::forward_as_tuple(layout, generator));
+    // The tuple slot is not something that is already in use.
+    EXPECT_TRUE(result.second);
+    InsertTuple(result.first->second, tested, layout, slot);
+    return *(result.first);
+  }
+
  protected:
   void SetUp() override {
     raw_block_ = block_store_.Get();
@@ -75,20 +168,20 @@ TEST_F(TupleAccessStrategyTests, SimpleInsertTest) {
         std::uniform_int_distribution<uint32_t>(1, layout.num_slots_)(generator);
     num_inserts = std::min(num_inserts, max_inserts);
 
-    std::unordered_map<storage::TupleSlot, testutil::FakeRawTuple> tuples;
+    std::unordered_map<storage::TupleSlot, FakeRawTuple> tuples;
 
     for (uint32_t j = 0; j < num_inserts; j++)
-      testutil::TryInsertFakeTuple(layout,
-                                   tested,
-                                   raw_block_,
-                                   tuples,
-                                   generator);
+      TryInsertFakeTuple(layout,
+                         tested,
+                         raw_block_,
+                         tuples,
+                         generator);
     // Check that all inserted tuples are equal to their expected values
     for (auto &entry : tuples)
-      testutil::CheckTupleEqual(entry.second,
-                                tested,
-                                layout,
-                                entry.first);
+      CheckTupleEqual(entry.second,
+                      tested,
+                      layout,
+                      entry.first);
   }
 }
 
@@ -155,27 +248,27 @@ TEST_F(TupleAccessStrategyTests, ConcurrentInsertTest) {
     storage::InitializeRawBlock(raw_block_, layout, VALUE_OF(layout_version_t, 0u));
     storage::TupleAccessStrategy tested(layout);
 
-    std::vector<std::unordered_map<storage::TupleSlot, testutil::FakeRawTuple>>
+    std::vector<std::unordered_map<storage::TupleSlot, FakeRawTuple>>
         tuples(num_threads);
 
     auto workload = [&](uint32_t id) {
       std::default_random_engine thread_generator(id);
       uint32_t work = std::min(layout.num_slots_, max_work);
       for (uint32_t j = 0; j < work / num_threads; j++)
-        testutil::TryInsertFakeTuple(layout,
-                                     tested,
-                                     raw_block_,
-                                     tuples[id],
-                                     thread_generator);
+        TryInsertFakeTuple(layout,
+                           tested,
+                           raw_block_,
+                           tuples[id],
+                           thread_generator);
     };
 
     testutil::RunThreadsUntilFinish(num_threads, workload);
     for (auto &thread_tuples : tuples)
       for (auto &entry : thread_tuples)
-        testutil::CheckTupleEqual(entry.second,
-                                  tested,
-                                  layout,
-                                  entry.first);
+        CheckTupleEqual(entry.second,
+                        tested,
+                        layout,
+                        entry.first);
   }
 }
 
@@ -201,17 +294,17 @@ TEST_F(TupleAccessStrategyTests, ConcurrentInsertDeleteTest) {
     storage::InitializeRawBlock(raw_block_, layout, VALUE_OF(layout_version_t, 0u));
     storage::TupleAccessStrategy tested(layout);
     std::vector<std::vector<storage::TupleSlot>> slots(num_threads);
-    std::vector<std::unordered_map<storage::TupleSlot, testutil::FakeRawTuple>>
+    std::vector<std::unordered_map<storage::TupleSlot, FakeRawTuple>>
         tuples(num_threads);
 
     auto workload = [&](uint32_t id) {
       std::default_random_engine thread_generator(id);
       auto insert = [&] {
-        auto &res = testutil::TryInsertFakeTuple(layout,
-                                                 tested,
-                                                 raw_block_,
-                                                 tuples[id],
-                                                 thread_generator);
+        auto &res = TryInsertFakeTuple(layout,
+                                       tested,
+                                       raw_block_,
+                                       tuples[id],
+                                       thread_generator);
         // log offset so we can pick random deletes
         slots[id].push_back(res.first);
       };
@@ -233,10 +326,10 @@ TEST_F(TupleAccessStrategyTests, ConcurrentInsertDeleteTest) {
     testutil::RunThreadsUntilFinish(num_threads, workload);
     for (auto &thread_tuples : tuples)
       for (auto &entry : thread_tuples)
-        testutil::CheckTupleEqual(entry.second,
-                                  tested,
-                                  layout,
-                                  entry.first);
+        CheckTupleEqual(entry.second,
+                        tested,
+                        layout,
+                        entry.first);
   }
 }
 }
